+++ src/main/java/MathOperation.java
@@ -1,10 +1,10 @@
 public class MathOperation {
 
-    private int number1;
-    private int number2;
+    private double number1;
+    private double number2;
     private String mathSymbol;
 
-    public MathOperation(int number1, int number2, String mathSymbol) {
+    public MathOperation(double number1, double number2, String mathSymbol) {
         this.number1 = number1;
         this.number2 = number2;
         this.mathSymbol = mathSymbol;
@@ -15,19 +15,19 @@ public class MathOperation {
         return number1 + " " + mathSymbol + " " + number2;
     }
 
-    public int getNumber1() {
+    public double getNumber1() {
         return number1;
     }
 
-    public void setNumber1(int number1) {
+    public void setNumber1(double number1) {
         this.number1 = number1;
     }
 
-    public int getNumber2() {
+    public double getNumber2() {
         return number2;
     }
 
-    public void setNumber2(int number2) {
+    public void setNumber2(double number2) {
         this.number2 = number2;
     }
